Bring back for saving usernames as well
While this also reintroduces that a password field is used for entering, this
is better than asking for the username every time, which is a regression
compared to 5.18.

Using a proper username input field would need changes in kwidgetsaddons or
some refactoring to ask for username and password at once.

Drop use of wallet->readEntry - it's custom binary data, so the content is
entirely defined by whoever wrote it. We don't use writeEntry here, so using
readEntry is invalud.